backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
python-multipart
pillow
numpy
scikit-learn
face_recognition
python-dotenv
